requirements.txt
requests
pyyaml
python-gnupg
